feat: drum pad cap 8→64 with MPC-style banking (item 24)
DrumMachine.max_pads grows 8->64 with lazy per-pad Sampler allocation
(pads: [64]?Sampler, materialized on first :load-pad rather than eager
— a 64-pad eager allocation would cost ~7.4MB of default-clip buffers
most projects never touch). The drum grid and sampler editor page pads
in banks of 8 (J/K jumps a bank, 1-8 selects within the current bank)
rather than scrolling, keeping the grid readable at any terminal size.

Two hard blockers surfaced during implementation, not caught by the
original scoping: the pad-encoded param id (`pad << 4 | param`) only
addressed 15 pads in a u8, widened to u16 across Event/Command; and
std.json's exact-length requirement for fixed arrays, which would have
broken every pre-existing save the moment max_pads changed — converted
DrumSnap/VariantSnap/ClipSnap's pad-indexed fields to slices. That's a
genuine semantic version bump (v10->v11): PadSnap gains `used` so an
absent pad can stay unmaterialized, version-gated against pre-v11
files (which had no such concept) rather than inferred from length.

+++ src/audio/engine.zig
@@ -46,8 +46,9 @@ pub const Command = union(enum) {
     cc: struct { track: u16, cc: u7, value: u7 },
     pitch_bend: struct { track: u16, bend: i16 },
     /// Nudge synth editor parameter `id` by `steps` on track `track`. Applied
-    /// on the audio thread so editor edits don't race the block reader.
-    set_track_param: struct { track: u16, id: u8, steps: i32 },
+    /// on the audio thread so editor edits don't race the block reader. u16,
+    /// not u8 — see dsp/device.zig's Event.set_param doc comment.
+    set_track_param: struct { track: u16, id: u16, steps: i32 },
     /// Which group (if any) `track` submixes through before the master bus.
     /// `null` routes straight to master, same as before grouping existed.
     set_track_group: struct { track: u16, group: ?u8 },
